4th assigement done
a)Add the interface Transportablewith the following method:
•public voidacceptVehicle(Vehicle vehicle)

b)Upgrade the class Vehicleby adding the loadFoodItem(FoodItem[] foodItems)method.

c)The  classes Warehouseand Storemust  implement  the Transportableinterface.The  method acceptVehiclein the Warehouseclass must loadthe accepted vehicle with food items that are currently in the warehouse. Alert: loading of food items varies iftheaccepted vehicle is a Van or a Truck (hint: use operatorinstanceof). Onto a truck,we can load a whole array of food items, while loading ontoa van is doneone food item at a time. The method acceptVehiclein the class Storeunloadsall food items from the vehicle.

d)While loading food items onto vehicles, many exceptions can occur. If a food item that we want to load onto a vehicle exceeds the maximal capacity of the vehicle(no more free spaces in the array), throw an exception named CapacityExceededException. If a food item that we want to load  onto  a  vehicleexceeds  the  volume limit  of  the  vehicle,  throw  an  exception  named VolumeExceededException. Alert:  if anexception  occurs  while  loading  an  array  of  fooditems onto a truck, unload all food items from the truck.

e)In the event that we want to load a food item with different food item type that van can accept, throw an exception named FoodItemTypeException.f)Throw  an  exception java.lang.IllegalArgumentException when  an  invalid  attribute  value  is entered, e.g., negative volume, maxWeight,and averageSpeed.Upgrade  the  class Launcherwritten  within  the third assignment  with  newly  added  concepts  and functionalities in order to demonstrate their functioning.